introduce 'helpers.intent'
This module should be used when a plugin wants to perform some operation
which can be done in various ways.

For example, a LocationList could intend to open a file but the opening
could open a new editor, or look for an open editor. So it should create
an intent and it will be handled. Intents are listened to, and when a
handler takes care of it, the intent stops being processed.