connector: caller stack info can be offset
connector decorators find the decorator caller in the stack in order to
ease de-registration.
However, if a plugin provides other helper decorators based on a
connector decorator (like helpers.intent), the identified caller would
be the helper plugin, not the end caller.

The stack offset can now be shifted in order to go a little upper in the
stack.