eye.widgets.editor: fix deadlock with isSignalConnected?
disconnectNotify() is called with a lock held and it isSignalConnected()
might hold the same lock too (not a recursive lock), causing a deadlock.

A solution is to count by ourselves the number of listeners of the
sciModified() signal. When it gets to 0, we don't need anymore
SCN_MODIFIED.

Also, it should have been a disconnect() call, not a connect().